@@ -3,7 +3,9 @@
   "type": "library",
   "license": "proprietary",
   "description": "General code for Calendar items",
-  "keywords": ["calendar"],
+  "keywords": [
+    "calendar"
+  ],
   "homepage": "https://itea3.org",
   "authors": [
     {
@@ -13,18 +15,18 @@
   ],
   "minimum-stability": "dev",
   "require": {
-    "php":">5.5",
+    "php": ">=5.6",
     "zendframework/zendframework": "~2.5"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.2.*",
-    "pdepend/pdepend": "2.0.*",
-    "phpmd/phpmd": "2.0.*",
-    "sebastian/phpcpd": "*",
-    "squizlabs/php_codesniffer": "2.*",
-    "fabpot/php-cs-fixer": "0.5.7.",
-    "jakub-onderka/php-parallel-lint": "0.*",
-    "jakub-onderka/php-console-highlighter": "0.*"
+    "phpunit/phpunit": "@stable",
+    "pdepend/pdepend": "@stable",
+    "phpmd/phpmd": "@stable",
+    "sebastian/phpcpd": "@stable",
+    "squizlabs/php_codesniffer": "@stable",
+    "fabpot/php-cs-fixer": "@stable",
+    "jakub-onderka/php-parallel-lint": "@stable",
+    "jakub-onderka/php-console-highlighter": "@stable"
   },
   "autoload": {
     "psr-4": {
@@ -32,10 +34,13 @@
       "CalendarTest\\": "tests/CalendarTest/"
     }
   },
+  "config": {
+    "secure-http": false
+  },
   "extra": {
     "branch-alias": {
-      "dev-master": "1.1-dev",
-      "dev-develop": "1.2-dev"
+      "dev-master": "2.0-dev",
+      "dev-develop": "2.1-dev"
     }
   }
 }
