@@ -1,23 +1,56 @@
 <?php
 // Generated by ZF2's ./bin/classmap_generator.php
 return array(
-    'Calendar\Controller\CalendarController'        => __DIR__ . '/src/Calendar/Controller/CalendarController.php',
-    'Calendar\Controller\CalendarManagerController' => __DIR__ . '/src/Calendar/Controller/CalendarManagerController.php',
-    'Calendar\Entity\EntityAbstract'                => __DIR__ . '/src/Calendar/Entity/EntityAbstract.php',
-    'Calendar\Entity\EntityInterface'               => __DIR__ . '/src/Calendar/Entity/EntityInterface.php',
-    'Calendar\Form\CreateObject'                    => __DIR__ . '/src/Calendar/Form/CreateObject.php',
-    'Calendar\Form\FilterCreateObject'              => __DIR__ . '/src/Calendar/Form/FilterCreateObject.php',
-    'Calendar\Form\ObjectFieldset'                  => __DIR__ . '/src/Calendar/Form/ObjectFieldset.php',
-    'Calendar\Module'                               => __DIR__ . '/src/Calendar/Module.php',
-    'Calendar\Service\CalendarService'              => __DIR__ . '/src/Calendar/Service/CalendarService.php',
-    'Calendar\Service\FormService'                  => __DIR__ . '/src/Calendar/Service/FormService.php',
-    'Calendar\Service\FormServiceAwareInterface'    => __DIR__ . '/src/Calendar/Service/FormServiceAwareInterface.php',
-    'Calendar\Service\ServiceAbstract'              => __DIR__ . '/src/Calendar/Service/ServiceAbstract.php',
-    'Calendar\Service\ServiceInterface'             => __DIR__ . '/src/Calendar/Service/ServiceInterface.php',
-    'Calendar\Version\Version'                      => __DIR__ . '/src/Calendar/Version/Version.php',
-    'Calendar\View\Helper\CalendarLink'             => __DIR__ . '/src/Calendar/View/Helper/CalendarLink.php',
-    'Calendar\Service\ModuleOptionAwareInterface'   => __DIR__ . '/src/Calendar/Service/ModuleOptionAwareInterface.php',
-    'ModuleOptionAwareInterface'                    => __DIR__ . '/src/Calendar/Controller/ControllerInitializer.php',
-    'Calendar\Options\CalendarOptionsInterface'     => __DIR__ . '/src/Calendar/Options/CalendarOptionsInterface.php',
-    'Calendar\Options\ModuleOptions'                => __DIR__ . '/src/Calendar/Options/ModuleOptions.php',
+    'Calendar\Repository\Calendar'                                   => __DIR__ . '/src/Calendar/Repository/Calendar.php',
+    'Calendar\Repository\Contact'                                    => __DIR__ . '/src/Calendar/Repository/Contact.php',
+    'Calendar\Options\CalendarOptionsInterface'                      => __DIR__ . '/src/Calendar/Options/CalendarOptionsInterface.php',
+    'Calendar\Options\ModuleOptions'                                 => __DIR__ . '/src/Calendar/Options/ModuleOptions.php',
+    'Calendar\Module'                                                => __DIR__ . '/src/Calendar/Module.php',
+    'Calendar\Factory\ModuleOptionsFactory'                          => __DIR__ . '/src/Calendar/Factory/ModuleOptionsFactory.php',
+    'Calendar\Factory\FormServiceFactory'                            => __DIR__ . '/src/Calendar/Factory/FormServiceFactory.php',
+    'Calendar\Factory\CalendarServiceFactory'                        => __DIR__ . '/src/Calendar/Factory/CalendarServiceFactory.php',
+    'Calendar\Entity\DocumentObject'                                 => __DIR__ . '/src/Calendar/Entity/DocumentObject.php',
+    'Calendar\Entity\Calendar'                                       => __DIR__ . '/src/Calendar/Entity/Calendar.php',
+    'Calendar\Entity\ScheduleContact'                                => __DIR__ . '/src/Calendar/Entity/ScheduleContact.php',
+    'Calendar\Entity\Contact'                                        => __DIR__ . '/src/Calendar/Entity/Contact.php',
+    'Calendar\Entity\Type'                                           => __DIR__ . '/src/Calendar/Entity/Type.php',
+    'Calendar\Entity\ContactStatus'                                  => __DIR__ . '/src/Calendar/Entity/ContactStatus.php',
+    'Calendar\Entity\EntityAbstract'                                 => __DIR__ . '/src/Calendar/Entity/EntityAbstract.php',
+    'Calendar\Entity\Document'                                       => __DIR__ . '/src/Calendar/Entity/Document.php',
+    'Calendar\Entity\EntityInterface'                                => __DIR__ . '/src/Calendar/Entity/EntityInterface.php',
+    'Calendar\Entity\Schedule'                                       => __DIR__ . '/src/Calendar/Entity/Schedule.php',
+    'Calendar\Entity\ContactRole'                                    => __DIR__ . '/src/Calendar/Entity/ContactRole.php',
+    'Calendar\Service\FormService'                                   => __DIR__ . '/src/Calendar/Service/FormService.php',
+    'Calendar\Service\ServiceInterface'                              => __DIR__ . '/src/Calendar/Service/ServiceInterface.php',
+    'Calendar\Service\CalendarService'                               => __DIR__ . '/src/Calendar/Service/CalendarService.php',
+    'Calendar\Service\ServiceAbstract'                               => __DIR__ . '/src/Calendar/Service/ServiceAbstract.php',
+    'Calendar\View\Helper\CalendarServiceProxy'                      => __DIR__ . '/src/Calendar/View/Helper/CalendarServiceProxy.php',
+    'Calendar\View\Helper\LinkAbstract'                              => __DIR__ . '/src/Calendar/View/Helper/LinkAbstract.php',
+    'Calendar\View\Helper\DocumentLink'                              => __DIR__ . '/src/Calendar/View/Helper/DocumentLink.php',
+    'Calendar\View\Helper\CalendarLink'                              => __DIR__ . '/src/Calendar/View/Helper/CalendarLink.php',
+    'Calendar\View\Helper\CalendarHandler'                           => __DIR__ . '/src/Calendar/View/Helper/CalendarHandler.php',
+    'Calendar\Version\Version'                                       => __DIR__ . '/src/Calendar/Version/Version.php',
+    'Calendar\Acl\Factory\AssertionInvokableAbstractFactory'         => __DIR__ . '/src/Calendar/Acl/Factory/AssertionInvokableAbstractFactory.php',
+    'Calendar\Acl\Assertion\AssertionAbstract'                       => __DIR__ . '/src/Calendar/Acl/Assertion/AssertionAbstract.php',
+    'Calendar\Acl\Assertion\Calendar'                                => __DIR__ . '/src/Calendar/Acl/Assertion/Calendar.php',
+    'Calendar\Acl\Assertion\Contact'                                 => __DIR__ . '/src/Calendar/Acl/Assertion/Contact.php',
+    'Calendar\Acl\Assertion\Document'                                => __DIR__ . '/src/Calendar/Acl/Assertion/Document.php',
+    'Calendar\Controller\CalendarAbstractController'                 => __DIR__ . '/src/Calendar/Controller/CalendarAbstractController.php',
+    'Calendar\Controller\CalendarCommunityController'                => __DIR__ . '/src/Calendar/Controller/CalendarCommunityController.php',
+    'Calendar\Controller\Factory\ControllerInvokableAbstractFactory' => __DIR__ . '/src/Calendar/Controller/Factory/ControllerInvokableAbstractFactory.php',
+    'Calendar\Controller\CalendarController'                         => __DIR__ . '/src/Calendar/Controller/CalendarController.php',
+    'Calendar\Controller\CalendarDocumentController'                 => __DIR__ . '/src/Calendar/Controller/CalendarDocumentController.php',
+    'Calendar\Controller\CalendarManagerController'                  => __DIR__ . '/src/Calendar/Controller/CalendarManagerController.php',
+    'Calendar\Controller\Plugin\RenderReviewCalendar'                => __DIR__ . '/src/Calendar/Controller/Plugin/RenderReviewCalendar.php',
+    'Calendar\Controller\Plugin\CalendarPdf'                         => __DIR__ . '/src/Calendar/Controller/Plugin/CalendarPdf.php',
+    'Calendar\Controller\Plugin\RenderCalendarContactList'           => __DIR__ . '/src/Calendar/Controller/Plugin/RenderCalendarContactList.php',
+    'Calendar\Navigation\Invokable\DocumentLabel'                    => __DIR__ . '/src/Calendar/Navigation/Invokable/DocumentLabel.php',
+    'Calendar\Navigation\Invokable\CalendarLabel'                    => __DIR__ . '/src/Calendar/Navigation/Invokable/CalendarLabel.php',
+    'Calendar\Form\CalendarContacts'                                 => __DIR__ . '/src/Calendar/Form/CalendarContacts.php',
+    'Calendar\Form\CreateCalendarDocument'                           => __DIR__ . '/src/Calendar/Form/CreateCalendarDocument.php',
+    'Calendar\Form\FilterCreateObject'                               => __DIR__ . '/src/Calendar/Form/FilterCreateObject.php',
+    'Calendar\Form\SendMessage'                                      => __DIR__ . '/src/Calendar/Form/SendMessage.php',
+    'Calendar\Form\ObjectFieldset'                                   => __DIR__ . '/src/Calendar/Form/ObjectFieldset.php',
+    'Calendar\Form\SelectAttendee'                                   => __DIR__ . '/src/Calendar/Form/SelectAttendee.php',
+    'Calendar\Form\CreateObject'                                     => __DIR__ . '/src/Calendar/Form/CreateObject.php',
 );
