@@ -5,25 +5,22 @@
  * @category   Project
  *
  * @author     Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright  2004-2014 ITEA Office
- * @license    http://debranova.org/license.txt proprietary
+ * @copyright  2004-2015 ITEA Office
+ * @license    https://itea3.org/license.txt proprietary
  *
- * @link       http://debranova.org
+ * @link       https://itea3.org
  */
 
 namespace Calendar\Acl\Assertion;
 
 use Admin\Entity\Access;
 use Admin\Service\AdminService;
-use Admin\Service\AdminServiceAwareInterface;
 use Calendar\Service\CalendarService;
-use Calendar\Service\CalendarServiceAwareInterface;
+use Contact\Entity\Contact;
 use Contact\Service\ContactService;
-use Contact\Service\ContactServiceAwareInterface;
 use Doctrine\ORM\PersistentCollection;
 use Zend\Mvc\Router\RouteMatch;
 use Zend\Permissions\Acl\Assertion\AssertionInterface;
-use Zend\ServiceManager\ServiceLocatorAwareInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
 
 /**
@@ -32,17 +29,12 @@ use Zend\ServiceManager\ServiceLocatorInterface;
  * @category   Calendar
  *
  * @author     Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright  2004-2014 ITEA Office
- * @license    http://debranova.org/license.txt proprietary
+ * @copyright  2004-2015 ITEA Office
+ * @license    https://itea3.org/license.txt proprietary
  *
- * @link       http://debranova.org
+ * @link       https://itea3.org
  */
-abstract class AssertionAbstract implements
-    AssertionInterface,
-    ServiceLocatorAwareInterface,
-    AdminServiceAwareInterface,
-    ContactServiceAwareInterface,
-    CalendarServiceAwareInterface
+abstract class AssertionAbstract implements AssertionInterface
 {
     /**
      * @var ServiceLocatorInterface
@@ -52,6 +44,10 @@ abstract class AssertionAbstract implements
      * @var ContactService
      */
     protected $contactService;
+    /**
+     * @var Contact
+     */
+    protected $contact;
     /**
      * @var AdminService
      */
@@ -74,88 +70,11 @@ abstract class AssertionAbstract implements
     }
 
     /**
-     * Get the service locator.
-     *
-     * @return ServiceLocatorInterface
-     */
-    public function getServiceLocator()
-    {
-        return $this->serviceLocator;
-    }
-
-    /**
-     * Set the service locator.
-     *
-     * @param ServiceLocatorInterface $serviceLocator
-     *
-     * @return AssertionAbstract
-     */
-    public function setServiceLocator(ServiceLocatorInterface $serviceLocator)
-    {
-        $this->serviceLocator = $serviceLocator;
-
-        return $this;
-    }
-
-    /**
      * @return bool
      */
     public function hasContact()
     {
-        return !$this->getContactService()->isEmpty();
-    }
-
-    /**
-     * @return ContactService
-     */
-    public function getContactService()
-    {
-        if ($this->contactService->isEmpty() && $this->getServiceLocator()->get('zfcuser_auth_service')->hasIdentity()
-        ) {
-            $this->contactService->setContact(
-                $this->getServiceLocator()->get('zfcuser_auth_service')->getIdentity()
-            );
-        }
-
-        return $this->contactService;
-    }
-
-    /**
-     * The contact service.
-     *
-     * @param ContactService $contactService
-     *
-     * @return $this;
-     */
-    public function setContactService(ContactService $contactService)
-    {
-        $this->contactService = $contactService;
-
-        return $this;
-    }
-
-    /**
-     * Get calendar service.
-     *
-     * @return CalendarService
-     */
-    public function getCalendarService()
-    {
-        return $this->calendarService;
-    }
-
-    /**
-     * The calendar service.
-     *
-     * @param CalendarService $calendarService
-     *
-     * @return $this
-     */
-    public function setCalendarService(CalendarService $calendarService)
-    {
-        $this->calendarService = $calendarService;
-
-        return $this;
+        return !$this->getContact()->isEmpty();
     }
 
     /**
@@ -179,9 +98,8 @@ abstract class AssertionAbstract implements
             if ($this->hasContact()) {
                 if (in_array(
                     strtolower($accessRole->getAccess()),
-                    $this->getAdminService()->findAccessRolesByContactAsArray($this->getContactService()->getContact())
-                )
-                ) {
+                    $this->getAdminService()->findAccessRolesByContactAsArray($this->getContact())
+                )) {
                     return true;
                 }
             }
@@ -214,11 +132,43 @@ abstract class AssertionAbstract implements
      */
     public function getAccessRoles()
     {
-        if (empty($this->accessRoles) && !$this->getContactService()->isEmpty()) {
-            $this->accessRoles = $this->getAdminService()->findAccessRolesByContactAsArray($this->getContactService()->getContact());
+        if (empty($this->accessRoles) && $this->hasContact()) {
+            $this->accessRoles = $this->getAdminService()->findAccessRolesByContactAsArray($this->getContact());
         }
 
         return $this->accessRoles;
+    }
+
+    /**
+     * @return ServiceLocatorInterface
+     */
+    public function getServiceLocator()
+    {
+        return $this->serviceLocator;
+    }
+
+    /**
+     * @param ServiceLocatorInterface $serviceLocator
+     */
+    public function setServiceLocator($serviceLocator)
+    {
+        $this->serviceLocator = $serviceLocator;
+    }
+
+    /**
+     * @return ContactService
+     */
+    public function getContactService()
+    {
+        return $this->contactService;
+    }
+
+    /**
+     * @param ContactService $contactService
+     */
+    public function setContactService($contactService)
+    {
+        $this->contactService = $contactService;
     }
 
     /**
@@ -231,12 +181,48 @@ abstract class AssertionAbstract implements
 
     /**
      * @param AdminService $adminService
+     */
+    public function setAdminService($adminService)
+    {
+        $this->adminService = $adminService;
+    }
+
+    /**
+     * @return CalendarService
+     */
+    public function getCalendarService()
+    {
+        return $this->calendarService;
+    }
+
+    /**
+     * @param CalendarService $calendarService
+     */
+    public function setCalendarService($calendarService)
+    {
+        $this->calendarService = $calendarService;
+    }
+
+    /**
+     * @return Contact
+     */
+    public function getContact()
+    {
+        if (is_null($this->contact)) {
+            $this->contact = new Contact();
+        }
+
+        return $this->contact;
+    }
+
+    /**
+     * @param Contact $contact
      *
      * @return AssertionAbstract
      */
-    public function setAdminService(AdminService $adminService)
+    public function setContact($contact)
     {
-        $this->adminService = $adminService;
+        $this->contact = $contact;
 
         return $this;
     }
