@@ -1,11 +1,11 @@
 <?php
 /**
- * Debranova copyright message placeholder.
+ * ITEA Office copyright message placeholder.
  *
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2014 Debranova
+ * @copyright Copyright (c) 2004-2015 ITEA Office (https://itea3.org)
  */
 
 namespace Calendar\Acl\Assertion;
@@ -32,20 +32,25 @@ class Document extends AssertionAbstract
      *
      * @return bool
      */
-    public function assert(Acl $acl, RoleInterface $role = null, ResourceInterface $resource = null, $privilege = null)
-    {
+    public function assert(
+        Acl $acl,
+        RoleInterface $role = null,
+        ResourceInterface $resource = null,
+        $privilege = null
+    ) {
         if (!$resource instanceof DocumentEntity) {
             /*
              * We are coming via the router, so we need to build up the information via the  routeMatch
              * The id and privilege are important
              */
-            $documentId = (int) $this->getRouteMatch()->getParam('id');
+            $documentId = (int)$this->getRouteMatch()->getParam('id');
             $privilege = $this->getRouteMatch()->getParam('privilege');
             /*
              * Check if a Contact has access to a meeting. We need to build the meeting first
              */
             $resource = $this->getCalendarService()->findEntityById('Document', $documentId);
         }
+
 
         /*
          * No document was found, so return true because we do not now anything about the access
@@ -60,9 +65,11 @@ class Document extends AssertionAbstract
         switch ($privilege) {
             case 'document-community':
             case 'download':
-                return $this->getCalendarService()->canViewCalendar($this->getContactService()->getContact());
+                return $this->getCalendarService()->canViewCalendar($this->getContact());
             case 'edit-community':
-                if ($this->getContactService()->hasPermit('edit', $resource->getCalendar())) {
+                if ($this->getContactService()
+                    ->contactHasPermit($this->getContact(), 'edit', $resource->getCalendar())
+                ) {
                     return true;
                 }
 
@@ -70,7 +77,8 @@ class Document extends AssertionAbstract
                  * The project leader also has rights to invite users
                  */
                 if (!is_null($resource->getCalendar()->getProjectCalendar())) {
-                    if ($this->getContactService()->hasPermit(
+                    if ($this->getContactService()->contactHasPermit(
+                        $this->getContact(),
                         'edit',
                         $resource->getCalendar()->getProjectCalendar()->getProject()
                     )
@@ -80,6 +88,9 @@ class Document extends AssertionAbstract
                 }
 
                 return $this->rolesHaveAccess([Access::ACCESS_OFFICE]);
+            case 'document-admin':
+                return $this->rolesHaveAccess([Access::ACCESS_OFFICE]);
+
         }
 
         return false;
