@@ -5,7 +5,7 @@
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2014 ITEA Office (http://itea3.org)
+ * @copyright Copyright (c) 2004-2015 ITEA Office (https://itea3.org)
  */
 
 namespace Calendar\Controller;
@@ -28,13 +28,10 @@ class CalendarDocumentController extends CalendarAbstractController
     public function downloadAction()
     {
         set_time_limit(0);
-        /*
-         * @var Document
+        /**
+         * @var  Document $document
          */
-        $document = $this->getCalendarService()->findEntityById(
-            'Document',
-            $this->getEvent()->getRouteMatch()->getParam('id')
-        );
+        $document = $this->getCalendarService()->findEntityById('Document', $this->params('id'));
         if (is_null($document) || sizeof($document->getObject()) === 0) {
             return $this->notFoundAction();
         }
@@ -45,16 +42,11 @@ class CalendarDocumentController extends CalendarAbstractController
         $object = $document->getObject()->first()->getObject();
         $response = $this->getResponse();
         $response->setContent(stream_get_contents($object));
-        $response->getHeaders()
-            ->addHeaderLine('Expires: '.gmdate('D, d M Y H:i:s \G\M\T', time() + 36000))
+        $response->getHeaders()->addHeaderLine('Expires: ' . gmdate('D, d M Y H:i:s \G\M\T', time() + 36000))
             ->addHeaderLine("Cache-Control: max-age=36000, must-revalidate")
-            ->addHeaderLine(
-                'Content-Disposition',
-                'attachment; filename="'.$document->parseFilename().'"'
-            )
-            ->addHeaderLine("Pragma: public")
-            ->addHeaderLine('Content-Type: '.$document->getContentType()->getContentType())
-            ->addHeaderLine('Content-Length: '.$document->getSize());
+            ->addHeaderLine('Content-Disposition', 'attachment; filename="' . $document->parseFileName() . '"')
+            ->addHeaderLine("Pragma: public")->addHeaderLine('Content-Type: ' . $document->getContentType()
+                    ->getContentType())->addHeaderLine('Content-Length: ' . $document->getSize());
 
         return $this->response;
     }
@@ -64,10 +56,7 @@ class CalendarDocumentController extends CalendarAbstractController
      */
     public function documentAction()
     {
-        $document = $this->getCalendarService()->findEntityById(
-            'document',
-            $this->getEvent()->getRouteMatch()->getParam('id')
-        );
+        $document = $this->getCalendarService()->findEntityById('document', $this->params('id'));
 
         if (is_null($document)) {
             return $this->notFoundAction();
@@ -81,19 +70,17 @@ class CalendarDocumentController extends CalendarAbstractController
      */
     public function editAction()
     {
-        $document = $this->getCalendarService()->findEntityById(
-            'document',
-            $this->getEvent()->getRouteMatch()->getParam('id')
-        );
+        /** @var Document $document */
+        $document = $this->getCalendarService()->findEntityById('document', $this->params('id'));
         if (is_null($document)) {
             return $this->notFoundAction();
         }
-        $entityManager = $this->getServiceLocator()->get('Doctrine\ORM\EntityManager');
+
         $data = array_merge_recursive(
             $this->getRequest()->getPost()->toArray(),
             $this->getRequest()->getFiles()->toArray()
         );
-        $form = new CreateCalendarDocument($entityManager);
+        $form = new CreateCalendarDocument($this->getEntityManager());
         $form->bind($document);
         $form->getInputFilter()->get('file')->setRequired(false);
         $form->setData($data);
@@ -106,16 +93,19 @@ class CalendarDocumentController extends CalendarAbstractController
              * Remove the file if delete is pressed
              */
             if (isset($data['delete'])) {
-                $this->flashMessenger()->setNamespace('success')->addMessage(
-                    sprintf(_("txt-calendar-document-%s-successfully-removed"), $document->parseFileName())
-                );
+                $this->flashMessenger()->setNamespace('success')
+                    ->addMessage(sprintf(
+                        $this->translate("txt-calendar-document-%s-successfully-removed"),
+                        $document->parseFileName()
+                    ));
                 $this->getCalendarService()->removeEntity($document);
 
-                return $this->redirect()->toRoute(
-                    'community/calendar/calendar',
-                    ['id' => $document->getCalendar()->getId()],
-                    ['fragment' => 'documents']
-                );
+                return $this->redirect()
+                    ->toRoute(
+                        'community/calendar/calendar',
+                        ['id' => $document->getCalendar()->getId()],
+                        ['fragment' => 'documents']
+                    );
             }
             /*
              * Handle when
@@ -129,9 +119,8 @@ class CalendarDocumentController extends CalendarAbstractController
                     $fileSizeValidator = new FilesSize(PHP_INT_MAX);
                     $fileSizeValidator->isValid($file);
                     $document->setSize($fileSizeValidator->size);
-                    $document->setContentType(
-                        $this->getGeneralService()->findContentTypeByContentTypeName($file['type'])
-                    );
+                    $document->setContentType($this->getGeneralService()
+                        ->findContentTypeByContentTypeName($file['type']));
                     /*
                      * Update the object
                      */
@@ -140,22 +129,19 @@ class CalendarDocumentController extends CalendarAbstractController
                     $this->getCalendarService()->updateEntity($documentObject);
                 }
                 $this->getCalendarService()->updateEntity($document);
-                $this->flashMessenger()->setNamespace('success')->addMessage(
-                    sprintf(_("txt-calendar-document-%s-successfully-updated"), $document->parseFileName())
-                );
+                $this->flashMessenger()->setNamespace('success')
+                    ->addMessage(sprintf(
+                        $this->translate("txt-calendar-document-%s-successfully-updated"),
+                        $document->parseFileName()
+                    ));
             }
 
-            return $this->redirect()->toRoute(
-                'community/calendar/document/document',
-                ['id' => $document->getId()]
-            );
+            return $this->redirect()->toRoute('community/calendar/document/document', ['id' => $document->getId()]);
         }
 
-        return new ViewModel(
-            [
-                'document' => $document,
-                'form'     => $form,
-            ]
-        );
+        return new ViewModel([
+            'document' => $document,
+            'form'     => $form,
+        ]);
     }
 }
