@@ -5,19 +5,24 @@
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2014 ITEA Office (http://itea3.org)
+ * @copyright Copyright (c) 2004-2015 ITEA Office (https://itea3.org)
  */
 
 namespace Calendar\Controller;
 
 use Calendar\Entity\Calendar;
+use Calendar\Entity\Contact;
+use Calendar\Entity\ContactRole;
 use Calendar\Entity\Document;
 use Calendar\Entity\DocumentObject;
+use Calendar\Form\CalendarContacts;
 use Calendar\Form\CreateCalendarDocument;
 use Calendar\Service\CalendarService;
+use Project\Service\ProjectService;
 use Zend\Paginator\Adapter\ArrayAdapter;
 use Zend\Paginator\Paginator;
 use Zend\Validator\File\FilesSize;
+use Zend\View\Model\JsonModel;
 use Zend\View\Model\ViewModel;
 
 /**
@@ -33,13 +38,10 @@ class CalendarManagerController extends CalendarAbstractController
     public function overviewAction()
     {
         $which = $this->getEvent()->getRouteMatch()->getParam('which', CalendarService::WHICH_UPCOMING);
-        $page = $this->getEvent()->getRouteMatch()->getParam('page', 1);
-        $year = $this->getEvent()->getRouteMatch()->getParam('year', date("Y"));
+        $page = $this->params('page', 1);
         $birthDays = $this->getContactService()->findContactsWithDateOfBirth();
-        $calendarItems = $this->getCalendarService()->findCalendarItems(
-            $which,
-            $this->zfcUserAuthentication()->getIdentity()
-        )->getResult();
+        $calendarItems = $this->getCalendarService()
+            ->findCalendarItems($which, $this->zfcUserAuthentication()->getIdentity())->getResult();
         $calender = [];
 
         $today = new \DateTime();
@@ -63,7 +65,7 @@ class CalendarManagerController extends CalendarAbstractController
 
             $calender[$index][] = [
                 'item' => sprintf(
-                    _("Birthday of %s (%s)"),
+                    $this->translate("Birthday of %s (%s)"),
                     $birthDay->getDisplayName(),
                     $birthDay->getDateOfBirth()->format("Y")
                 ),
@@ -95,13 +97,11 @@ class CalendarManagerController extends CalendarAbstractController
         $paginator->setPageRange(ceil($paginator->getTotalItemCount() / $paginator->getDefaultItemCountPerPage()));
         $whichValues = $this->getCalendarService()->getWhichValues();
 
-        return new ViewModel(
-            [
-                'which'       => $which,
-                'paginator'   => $paginator,
-                'whichValues' => $whichValues,
-            ]
-        );
+        return new ViewModel([
+            'which'       => $which,
+            'paginator'   => $paginator,
+            'whichValues' => $whichValues,
+        ]);
     }
 
     /**
@@ -111,33 +111,58 @@ class CalendarManagerController extends CalendarAbstractController
      */
     public function newAction()
     {
+        $projectService = null;
+        if (!is_null($this->params('project'))) {
+            $projectService = $this->getProjectService()->setProjectId($this->params('project'));
+            if ($projectService->isEmpty()) {
+                return $this->notFoundAction();
+            }
+        }
+
         $calendar = new Calendar();
         $data = $this->getRequest()->getPost()->toArray();
         $form = $this->getFormService()->prepare('calendar', $calendar, $data);
-        if ($this->getRequest()->isPost() && $form->isValid()) {
+        $form->remove('delete');
+        if ($this->getRequest()->isPost()) {
             /*
              * Return when cancel is pressed
              */
             if (isset($data['cancel'])) {
                 return $this->redirect()->toRoute('zfcadmin/calendar-manager/overview');
             }
-            /**
-             * @var $calendar Calendar
-             */
-            $calendar = $form->getData();
-            $calendar->setContact($this->zfcUserAuthentication()->getIdentity());
-            $calendar = $this->getCalendarService()->newEntity($calendar);
-            $this->flashMessenger()->setNamespace('success')->addMessage(
-                sprintf(
-                    $this->translate("txt-calendar-item-%s-has-been-created-successfully"),
-                    $calendar->getCalendar()
-                )
-            );
 
-            return $this->redirect()->toRoute('zfcadmin/calendar-manager/calendar', ['id' => $calendar->getId()]);
+            if ($form->isValid()) {
+                /**
+                 * @var $calendar Calendar
+                 */
+                $calendar = $form->getData();
+                $calendar->setContact($this->zfcUserAuthentication()->getIdentity());
+                $calendar = $this->getCalendarService()->newEntity($calendar);
+
+                /**
+                 * @var ProjectService|null $projectService
+                 */
+                if (!is_null($projectService)) {
+                    $projectCalendar = new \Project\Entity\Calendar\Calendar();
+                    $projectCalendar->setProject($projectService->getProject());
+                    $projectCalendar->setCalendar($calendar);
+                    $this->getProjectService()->updateEntity($projectCalendar);
+                }
+
+                $this->flashMessenger()->setNamespace('success')
+                    ->addMessage(sprintf(
+                        $this->translate("txt-calendar-item-%s-has-been-created-successfully"),
+                        $calendar->getCalendar()
+                    ));
+
+                return $this->redirect()->toRoute('zfcadmin/calendar-manager/calendar', ['id' => $calendar->getId()]);
+            }
         }
 
-        return new ViewModel(['form' => $form]);
+        return new ViewModel([
+            'form'           => $form,
+            'projectService' => $projectService
+        ]);
     }
 
     /**
@@ -147,9 +172,7 @@ class CalendarManagerController extends CalendarAbstractController
      */
     public function editAction()
     {
-        $calendarService = $this->getCalendarService()->setCalendarId(
-            $this->getEvent()->getRouteMatch()->getParam('id')
-        );
+        $calendarService = $this->getCalendarService()->setCalendarId($this->params('id'));
 
         $data = $this->getRequest()->getPost()->toArray();
 
@@ -159,10 +182,8 @@ class CalendarManagerController extends CalendarAbstractController
              * Return when cancel is pressed
              */
             if (isset($data['cancel'])) {
-                return $this->redirect()->toRoute(
-                    'zfcadmin/calendar-manager/calendar',
-                    ['id' => $calendarService->getCalendar()->getId()]
-                );
+                return $this->redirect()
+                    ->toRoute('zfcadmin/calendar-manager/calendar', ['id' => $calendarService->getCalendar()->getId()]);
             }
             /*
              * Return when cancel is pressed
@@ -170,12 +191,11 @@ class CalendarManagerController extends CalendarAbstractController
             if (isset($data['delete'])) {
                 $this->getCalendarService()->removeEntity($calendarService->getCalendar());
 
-                $this->flashMessenger()->setNamespace('success')->addMessage(
-                    sprintf(
+                $this->flashMessenger()->setNamespace('success')
+                    ->addMessage(sprintf(
                         $this->translate("txt-calendar-item-%s-has-been-deleted-successfully"),
                         $calendarService->getCalendar()->getCalendar()
-                    )
-                );
+                    ));
 
                 return $this->redirect()->toRoute('zfcadmin/calendar-manager/overview');
             }
@@ -187,12 +207,11 @@ class CalendarManagerController extends CalendarAbstractController
                 $calendar = $form->getData();
                 $calendar->setContact($this->zfcUserAuthentication()->getIdentity());
                 $calendar = $this->getCalendarService()->newEntity($calendar);
-                $this->flashMessenger()->setNamespace('success')->addMessage(
-                    sprintf(
+                $this->flashMessenger()->setNamespace('success')
+                    ->addMessage(sprintf(
                         $this->translate("txt-calendar-item-%s-has-been-updated-successfully"),
                         $calendar->getCalendar()
-                    )
-                );
+                    ));
 
                 return $this->redirect()->toRoute('zfcadmin/calendar-manager/calendar', ['id' => $calendar->getId()]);
             }
@@ -202,31 +221,153 @@ class CalendarManagerController extends CalendarAbstractController
     }
 
     /**
+     * Action to edit a calendar element.
+     *
+     * @return ViewModel
+     */
+    public function selectAttendeesAction()
+    {
+        $calendarService = $this->getCalendarService()->setCalendarId($this->params('id'));
+
+        $data = array_merge($this->getRequest()->getPost()->toArray());
+
+
+        $form = new CalendarContacts($this->getSelectionService());
+        $form->setData($data);
+
+
+        if ($this->getRequest()->isPost()) {
+            if (isset($data['cancel'])) {
+                return $this->redirect()
+                    ->toRoute('zfcadmin/calendar-manager/calendar', ['id' => $calendarService->getCalendar()->getId()]);
+            }
+
+
+            $this->getCalendarService()->updateCalendarContacts($calendarService->getCalendar(), $data);
+
+            return $this->redirect()
+                ->toRoute('zfcadmin/calendar-manager/calendar', ['id' => $calendarService->getCalendar()->getId()]);
+        }
+
+        return new ViewModel([
+            'calendarService' => $calendarService,
+            'form'            => $form
+        ]);
+    }
+
+    /**
+     * Action to edit a calendar element.
+     *
+     * @return ViewModel
+     */
+    public function setRolesAction()
+    {
+        $calendarService = $this->getCalendarService()->setCalendarId($this->params('id'));
+
+        $data = $this->getRequest()->getPost()->toArray();
+
+        $form = $this->getFormService()->prepare('calendar', $calendarService->getCalendar(), $data);
+        if ($this->getRequest()->isPost()) {
+            /*
+             * Return when cancel is pressed
+             */
+            if (isset($data['cancel'])) {
+                return $this->redirect()
+                    ->toRoute('zfcadmin/calendar-manager/calendar', ['id' => $calendarService->getCalendar()->getId()]);
+            }
+
+
+            if ($form->isValid()) {
+                $this->flashMessenger()->setNamespace('success')
+                    ->addMessage(sprintf(
+                        $this->translate("txt-calendar-item-%s-has-been-updated-successfully"),
+                        $calendarService->getCalendar()
+                    ));
+
+                return $this->redirect()
+                    ->toRoute('zfcadmin/calendar-manager/calendar', ['id' => $calendarService->getId()]);
+            }
+        }
+
+        return new ViewModel([
+            'form'            => $form,
+            'calendarService' => $calendarService
+        ]);
+    }
+
+    /**
+     * @return JsonModel
+     */
+    public function getRolesAction()
+    {
+        $calendarRoles = $this->getCalendarService()->findAll('contactRole');
+
+        $roles = [];
+        /**
+         * @var $role ContactRole
+         */
+        foreach ($calendarRoles as $role) {
+            $roles[] = [
+                'value' => $role->getId(),
+                'text'  => $role->getRole(),
+            ];
+        }
+
+        return new JsonModel($roles);
+    }
+
+    public function updateRoleAction()
+    {
+        $calendarContactId = $this->params()->fromPost('pk');
+        $roleId = $this->params()->fromPost('value');
+
+        /**
+         * @var $calendarContact Contact
+         */
+        $calendarContact = $this->getCalendarService()->findEntityById('contact', $calendarContactId);
+
+        if (is_null($calendarContact)) {
+            return $this->notFoundAction();
+        }
+        /**
+         * @var $role ContactRole
+         */
+        $role = $this->getCalendarService()->findEntityById('contactRole', $roleId);
+
+        if (is_null($role)) {
+            return $this->notFoundAction();
+        }
+
+        $calendarContact->setRole($role);
+        $this->getCalendarService()->updateEntity($calendarContact);
+
+        return new JsonModel();
+    }
+
+    /**
      * @return ViewModel
      */
     public function calendarAction()
     {
-        $calendarService = $this->getCalendarService()->setCalendarId(
-            $this->getEvent()->getRouteMatch()->getParam('id')
-        );
+        $calendarService = $this->getCalendarService()->setCalendarId($this->params('id'));
         if (is_null($calendarService->getCalendar()->getId())) {
             return $this->notFoundAction();
         }
-        $data = array_merge_recursive(
-            $this->getRequest()->getPost()->toArray(),
-            $this->getRequest()->getFiles()->toArray()
-        );
-        $entityManager = $this->getServiceLocator()->get('Doctrine\ORM\EntityManager');
-        $form = new CreateCalendarDocument($entityManager);
+        $data = array_merge($this->getRequest()->getPost()->toArray(), $this->getRequest()->getFiles()->toArray());
+
+        $form = new CreateCalendarDocument($this->getEntityManager());
         $form->bind(new Document());
         //Add the missing form fields
         $data['calendar'] = $calendarService->getCalendar()->getId();
         $form->setData($data);
         if ($this->getRequest()->isPost() && $form->isValid()) {
+            /**
+             * @var $document Document
+             */
             $document = $form->getData();
             $document->setCalendar($calendarService->getCalendar());
             $document->setContact($this->zfcUserAuthentication()->getIdentity());
-            /*
+            /**
              * Add the file
              */
             $file = $data['file'];
@@ -237,32 +378,31 @@ class CalendarManagerController extends CalendarAbstractController
             $documentObject = new DocumentObject();
             $documentObject->setDocument($document);
             $documentObject->setObject(file_get_contents($file['tmp_name']));
+
+            if (empty($data['document'])) {
+                $document->setDocument($file['name']);
+            }
+
             $this->getCalendarService()->updateEntity($documentObject);
-            $this->flashMessenger()->addInfoMessage(
-                sprintf(
+            $this->flashMessenger()
+                ->addInfoMessage(sprintf(
                     $this->translate("txt-calendar-document-%s-for-calendar-%s-has-successfully-been-uploaded"),
                     $document->getDocument(),
                     $calendarService->getCalendar()->getCalendar()
-                )
-            );
+                ));
 
             /*
              * Document uploaded
              */
 
-            return $this->redirect()->toRoute(
-                'zfcadmin/calendar-manager/calendar',
-                [
-                    'id' => $calendarService->getCalendar()->getId(),
-                ]
-            );
+            return $this->redirect()->toRoute('zfcadmin/calendar-manager/calendar', [
+                'id' => $calendarService->getCalendar()->getId(),
+            ]);
         }
 
-        return new ViewModel(
-            [
-                'calendarService' => $calendarService,
-                'form'            => $form,
-            ]
-        );
+        return new ViewModel([
+            'calendarService' => $calendarService,
+            'form'            => $form,
+        ]);
     }
 }
