@@ -5,7 +5,7 @@
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2014 ITEA Office (http://itea3.org)
+ * @copyright Copyright (c) 2004-2015 ITEA Office (https://itea3.org)
  */
 
 namespace Calendar\Form;
@@ -33,63 +33,50 @@ class CreateCalendarDocument extends Form
         $this->setAttribute('action', '');
         $this->setAttribute('class', 'form-horizontal');
         $this->setAttribute('id', 'create-document');
-        $this->add(
-            [
-                'type'       => 'Zend\Form\Element\Text',
-                'name'       => 'document',
-                'options'    => [
-                    'label'      => _("txt-document-name"),
-                    'help-block' => _("txt-document-name-explanation"),
-                ],
-                'attributes' => [
-                    'required'    => true,
-                    'class'       => 'form-control',
-                    'placeholder' => _("txt-please-give-a-calendar-document-name"),
-                ],
-            ]
-        );
-        $this->add(
-            [
-                'type'       => '\Zend\Form\Element\File',
-                'name'       => 'file',
-                'options'    => [
-                    "label"      => "txt-file",
-                    "help-block" => _("txt-file-requirements"),
-                ],
-                'attributes' => [
-                    'class' => 'form-control',
-                ],
-            ]
-        );
-        $this->add(
-            [
-                'type'       => 'Zend\Form\Element\Submit',
-                'name'       => 'submit',
-                'attributes' => [
-                    'class' => "btn btn-primary",
-                    'value' => _("txt-update"),
-                ],
-            ]
-        );
-        $this->add(
-            [
-                'type'       => 'Zend\Form\Element\Submit',
-                'name'       => 'delete',
-                'attributes' => [
-                    'class' => "btn btn-danger",
-                    'value' => _("txt-delete"),
-                ],
-            ]
-        );
-        $this->add(
-            [
-                'type'       => 'Zend\Form\Element\Submit',
-                'name'       => 'cancel',
-                'attributes' => [
-                    'class' => "btn btn-warning",
-                    'value' => _("txt-cancel"),
-                ],
-            ]
-        );
+        $this->add([
+            'type'       => 'Zend\Form\Element\Text',
+            'name'       => 'document',
+            'options'    => [
+                'label'      => _("txt-document-name"),
+                'help-block' => _("txt-document-name-explanation"),
+            ],
+            'attributes' => [
+                'class'       => 'form-control',
+                'placeholder' => _("txt-please-give-a-calendar-document-name"),
+            ],
+        ]);
+        $this->add([
+            'type'    => '\Zend\Form\Element\File',
+            'name'    => 'file',
+            'options' => [
+                "label"      => "txt-file",
+                'class'      => 'form-control',
+                "help-block" => _("txt-file-requirements"),
+            ],
+        ]);
+        $this->add([
+            'type'       => 'Zend\Form\Element\Submit',
+            'name'       => 'submit',
+            'attributes' => [
+                'class' => "btn btn-primary",
+                'value' => _("txt-update"),
+            ],
+        ]);
+        $this->add([
+            'type'       => 'Zend\Form\Element\Submit',
+            'name'       => 'delete',
+            'attributes' => [
+                'class' => "btn btn-danger",
+                'value' => _("txt-delete"),
+            ],
+        ]);
+        $this->add([
+            'type'       => 'Zend\Form\Element\Submit',
+            'name'       => 'cancel',
+            'attributes' => [
+                'class' => "btn btn-warning",
+                'value' => _("txt-cancel"),
+            ],
+        ]);
     }
 }
