@@ -5,7 +5,7 @@
  * @category   SoloDB
  *
  * @author     Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright  Copyright (c) 2004-2014 ITEA Office (http://itea3.org)
+ * @copyright  Copyright (c) 2004-2014 ITEA Office (https://itea3.org)
  *
  * @version    4.0
  */
@@ -14,28 +14,24 @@ namespace Calendar;
 
 use Calendar\Controller\Plugin\RenderCalendarContactList;
 use Calendar\Controller\Plugin\RenderReviewCalendar;
-use Zend\EventManager\EventInterface;
+use Calendar\Navigation;
 use Zend\ModuleManager\Feature;
-use Zend\Mvc\MvcEvent;
+use Zend\Mvc\Controller\PluginManager;
 
 /**
  * @author
  */
-class Module implements
-    Feature\AutoloaderProviderInterface,
-    Feature\ServiceProviderInterface,
-    Feature\ConfigProviderInterface,
-    Feature\BootstrapListenerInterface
+class Module implements Feature\AutoloaderProviderInterface, Feature\ServiceProviderInterface, Feature\ConfigProviderInterface
 {
     public function getAutoloaderConfig()
     {
         return [
             'Zend\Loader\ClassMapAutoloader' => [
-                __DIR__.'/../../autoload_classmap.php',
+                __DIR__ . '/../../autoload_classmap.php',
             ],
             'Zend\Loader\StandardAutoloader' => [
                 'namespaces' => [
-                    __NAMESPACE__ => __DIR__.'/../../src/'.__NAMESPACE__,
+                    __NAMESPACE__ => __DIR__ . '/../../src/' . __NAMESPACE__,
                 ],
             ],
         ];
@@ -46,7 +42,7 @@ class Module implements
      */
     public function getConfig()
     {
-        return include __DIR__.'/../../config/module.config.php';
+        return include __DIR__ . '/../../config/module.config.php';
     }
 
     /**
@@ -56,7 +52,7 @@ class Module implements
      */
     public function getServiceConfig()
     {
-        return include __DIR__.'/../../config/services.config.php';
+        return include __DIR__ . '/../../config/services.config.php';
     }
 
     /**
@@ -64,7 +60,7 @@ class Module implements
      */
     public function getViewHelperConfig()
     {
-        return include __DIR__.'/../../config/viewhelpers.config.php';
+        return include __DIR__ . '/../../config/viewhelpers.config.php';
     }
 
     /**
@@ -76,13 +72,14 @@ class Module implements
     {
         return [
             'factories' => [
-                'renderCalendarContactList' => function ($sm) {
-                    $renderCalendarContactList = new RenderCalendarContactList();
+                'renderCalendarContactList' => function (PluginManager $sm) {
+                    $renderCalendarContactList
+                        = new RenderCalendarContactList();
                     $renderCalendarContactList->setServiceLocator($sm->getServiceLocator());
 
                     return $renderCalendarContactList;
                 },
-                'renderReviewCalendar'      => function ($sm) {
+                'renderReviewCalendar'      => function (PluginManager $sm) {
                     $renderReviewCalendar = new RenderReviewCalendar();
                     $renderReviewCalendar->setServiceLocator($sm->getServiceLocator());
 
@@ -90,24 +87,5 @@ class Module implements
                 },
             ],
         ];
-    }
-
-    /**
-     * Listen to the bootstrap event.
-     *
-     * @param EventInterface $e
-     *
-     * @return array
-     */
-    public function onBootstrap(EventInterface $e)
-    {
-        $app = $e->getParam('application');
-        $em = $app->getEventManager();
-        $em->attach(
-            MvcEvent::EVENT_DISPATCH,
-            function ($event) {
-                $event->getApplication()->getServiceManager()->get('calendar_navigation_service')->update();
-            }
-        );
     }
 }
