@@ -5,21 +5,20 @@
  * @category  Calendar
  * @package   Service
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2014 ITEA Office (http://itea3.org)
+ * @copyright Copyright (c) 2004-2014 ITEA Office (https://itea3.org)
  */
 namespace Calendar\Service;
 
 use Calendar\Entity;
 use Calendar\Entity\Calendar;
 use Calendar\Entity\Contact as CalendarContact;
-use Calendar\Options\ModuleOptions;
 use Contact\Entity\Contact;
 use Project\Entity\Project;
 
 /**
  *
  */
-class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterface
+class CalendarService extends ServiceAbstract
 {
     /**
      * Constant to determine which affiliations must be taken from the database
@@ -33,10 +32,6 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      * @var Entity\Calendar
      */
     protected $calendar;
-    /**
-     * @var CalendarService
-     */
-    protected $calendarService;
 
     /**
      * @param int $id
@@ -65,37 +60,16 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      */
     public function findCalendarByDocRef($docRef)
     {
-        $calendar = $this->getEntityManager()->getRepository($this->getFullEntityName('Calendar'))->findOneBy(
-            [
-                'docRef' => $docRef,
-            ]
-        );
+        $calendar = $this->getEntityManager()->getRepository(Entity\Calendar::class)->findOneBy([
+            'docRef' => $docRef,
+        ]);
         if (is_null($calendar)) {
-            return;
+            return null;
         }
 
         return $calendar;
     }
 
-    /**
-     * @return ModuleOptions
-     */
-    public function getOptions()
-    {
-        return $this->options;
-    }
-
-    /**
-     * @param ModuleOptions $options
-     *
-     * @return ServiceAbstract
-     */
-    public function setOptions(\Calendar\Options\ModuleOptions $options)
-    {
-        $this->options = $options;
-
-        return $this;
-    }
 
     /**
      * @param Calendar $calendar
@@ -105,27 +79,87 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      */
     public function calendarHasContact(Calendar $calendar, Contact $contact)
     {
-        $calendarContact = $this->getEntityManager()
-            ->getRepository($this->getFullEntityName('Contact'))
-            ->findOneBy(
-                [
-                    'calendar' => $calendar,
-                    'contact'  => $contact,
-                ]
-            );
+        $calendarContact = $this->getEntityManager()->getRepository(CalendarContact::class)->findOneBy([
+            'calendar' => $calendar,
+            'contact'  => $contact,
+        ]);
 
         return !is_null($calendarContact);
     }
 
     /**
+     * @param array $data
+     * @param Calendar $calendar
+     *
+     * array (size=5)
+     * 'type' => string '2' (length=1)
+     * 'added' => string '20388' (length=5)
+     * 'removed' => string '' (length=0)
+     * 'sql' => string '' (length=0)
+     *
+     *
+     */
+    public function updateCalendarContacts(Calendar $calendar, array $data)
+    {
+        //Update the contacts
+        if (!empty($data['added'])) {
+            foreach (explode(',', $data['added']) as $contactId) {
+                $contact = $this->getContactService()->findEntityById('contact', $contactId);
+
+                $contactService = clone $this->getContactService();
+                $contactService->setContact($contact);
+
+                if (!$contactService->isEmpty()
+                    && !$this->calendarHasContact($calendar, $contact)
+                ) {
+                    $calendarContact = new CalendarContact();
+                    $calendarContact->setContact($contact);
+                    $calendarContact->setCalendar($calendar);
+
+                    /**
+                     * Add every new user as attendee
+                     *
+                     * @var $role Entity\ContactRole
+                     */
+                    $role = $this->findEntityById('contactRole', Entity\ContactRole::ROLE_ATTENDEE);
+                    $calendarContact->setRole($role);
+
+                    /**
+                     * Add every new user as attendee
+                     *
+                     * @var $status Entity\ContactStatus
+                     */
+                    $status = $this->findEntityById('contactStatus', Entity\ContactStatus::STATUS_TENTATIVE);
+                    $calendarContact->setStatus($status);
+
+                    $this->newEntity($calendarContact);
+                }
+            }
+        }
+
+        //Update the contacts
+        if (!empty($data['removed'])) {
+            foreach (explode(',', $data['removed']) as $contactId) {
+                foreach ($calendar->getCalendarContact() as $calendarContact) {
+                    if ($calendarContact->getContact()->getId() === (int)$contactId) {
+                        $this->removeEntity($calendarContact);
+                    }
+                }
+            }
+        }
+    }
+
+    /**
      * @param  string $which
      * @param  Contact $contact
+     *
      * @return CalendarContact[]
      */
-    public function findCalendarContactByContact($which = self::WHICH_UPCOMING, Contact $contact = null)
-    {
-        return $this->getEntityManager()
-            ->getRepository($this->getFullEntityName('Contact'))
+    public function findCalendarContactByContact(
+        $which = self::WHICH_UPCOMING,
+        Contact $contact = null
+    ) {
+        return $this->getEntityManager()->getRepository(CalendarContact::class)
             ->findCalendarContactByContact($which, $contact);
     }
 
@@ -135,23 +169,24 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      *
      * @return CalendarContact
      */
-    public function findCalendarContactByContactAndCalendar(Contact $contact, Calendar $calendar)
-    {
-        return $this->getEntityManager()
-            ->getRepository($this->getFullEntityName('Contact'))
+    public function findCalendarContactByContactAndCalendar(
+        Contact $contact,
+        Calendar $calendar
+    ) {
+        return $this->getEntityManager()->getRepository(CalendarContact::class)
             ->findCalendarContactByContactAndCalendar($contact, $calendar);
     }
 
     /**
      * @param Calendar $calendar
+     * @param int $status
      *
      * @return CalendarContact[]
      */
-    public function findCalendarContactsByCalendar(Calendar $calendar)
+    public function findCalendarContactsByCalendar(Calendar $calendar, $status = CalendarContact::STATUS_ALL)
     {
-        return $this->getEntityManager()
-            ->getRepository($this->getFullEntityName('Contact'))
-            ->findCalendarContactsByCalendar($calendar);
+        return $this->getEntityManager()->getRepository(CalendarContact::class)
+            ->findCalendarContactsByCalendar($calendar, $status);
     }
 
     /**
@@ -163,8 +198,7 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      */
     public function canViewCalendar(Contact $contact = null)
     {
-        return $this->getEntityManager()
-            ->getRepository($this->getFullEntityName('Calendar'))
+        return $this->getEntityManager()->getRepository(Entity\Calendar::class)
             ->canViewCalendar($this->getCalendar(), $contact);
     }
 
@@ -182,8 +216,7 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
         $year = null,
         $type = null
     ) {
-        return $this->getEntityManager()
-            ->getRepository($this->getFullEntityName('Calendar'))
+        return $this->getEntityManager()->getRepository(Entity\Calendar::class)
             ->findCalendarItems($which, true, $contact, $year, $type);
     }
 
@@ -200,12 +233,17 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
          * Add the calendar items from the project
          */
         foreach ($project->getProjectCalendar() as $calendarItem) {
-            if (!$onlyFinal || $calendarItem->getCalendar()->getFinal() === Calendar::FINAL_FINAL) {
-                $calendar[$calendarItem->getCalendar()->getId()] = $calendarItem->getCalendar();
+            if (!$onlyFinal
+                || $calendarItem->getCalendar()->getFinal() === Calendar::FINAL_FINAL
+            ) {
+                $calendar[$calendarItem->getCalendar()->getId()]
+                    = $calendarItem->getCalendar();
             }
         }
         foreach ($project->getCall()->getCalendar() as $calendarItem) {
-            if (!$onlyFinal || $calendarItem->getFinal() === Calendar::FINAL_FINAL) {
+            if (!$onlyFinal
+                || $calendarItem->getFinal() === Calendar::FINAL_FINAL
+            ) {
                 if ($calendarItem->getDateEnd() > new \DateTime()) {
                     $calendar[$calendarItem->getId()] = $calendarItem;
                 }
@@ -224,9 +262,7 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      */
     public function findLatestProjectCalendar(Project $project)
     {
-        return $this->getEntityManager()->getRepository(
-            $this->getFullEntityName('calendar')
-        )->findLatestProjectCalendar($project);
+        return $this->getEntityManager()->getRepository(Entity\Calendar::class)->findLatestProjectCalendar($project);
     }
 
     /**
@@ -237,8 +273,10 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      *
      * @return Calendar|null
      */
-    public function findNextProjectCalendar(Project $project, \DateTime $datetime)
-    {
+    public function findNextProjectCalendar(
+        Project $project,
+        \DateTime $datetime
+    ) {
         return $this->getEntityManager()->getRepository(Calendar::class)->findNextProjectCalendar($project, $datetime);
     }
 
@@ -250,23 +288,24 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      *
      * @return Calendar|null
      */
-    public function findPreviousProjectCalendar(Project $project, \DateTime $datetime)
-    {
-        return $this->getEntityManager()->getRepository(Calendar::class)->findPreviousProjectCalendar(
-            $project,
-            $datetime
-        );
+    public function findPreviousProjectCalendar(
+        Project $project,
+        \DateTime $datetime
+    ) {
+        return $this->getEntityManager()->getRepository(Calendar::class)
+            ->findPreviousProjectCalendar($project, $datetime);
     }
 
     /**
      * @param CalendarContact $calendarContact
      * @param                 $status
      */
-    public function updateContactStatus(CalendarContact $calendarContact, $status)
-    {
-        $calendarContact->setStatus(
-            $this->getEntityManager()->getReference($this->getFullEntityName('ContactStatus'), $status)
-        );
+    public function updateContactStatus(
+        CalendarContact $calendarContact,
+        $status
+    ) {
+        $calendarContact->setStatus($this->getEntityManager()
+            ->getReference($this->getFullEntityName('ContactStatus'), $status));
         $this->updateEntity($calendarContact);
     }
 
@@ -277,8 +316,7 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
      */
     public function findGeneralCalendarContactByCalendar(Calendar $calendar)
     {
-        return $this->getEntityManager()
-            ->getRepository($this->getFullEntityName('Contact'))
+        return $this->getEntityManager()->getRepository(CalendarContact::class)
             ->findGeneralCalendarContactByCalendar($calendar);
     }
 
@@ -295,25 +333,5 @@ class CalendarService extends ServiceAbstract implements ModuleOptionAwareInterf
             self::WHICH_PAST,
             self::WHICH_REVIEWS,
         ];
-    }
-
-    /**
-     * @param \Calendar\Entity\Calendar $calendar
-     *
-     * @return $this;
-     */
-    public function setCalendar($calendar)
-    {
-        $this->calendar = $calendar;
-
-        return $this;
-    }
-
-    /**
-     * @return \Calendar\Entity\Calendar
-     */
-    public function getCalendar()
-    {
-        return $this->calendar;
     }
 }
