@@ -6,20 +6,13 @@
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2014 ITEA Office (http://itea3.org)
+ * @copyright Copyright (c) 2004-2015 ITEA Office (https://itea3.org)
  */
 
 namespace Calendar\Service;
 
-use Zend\Form\Form;
-
 class FormService extends ServiceAbstract
 {
-    /**
-     * @var Form
-     */
-    protected $form;
-
     /**
      * @param null $className
      * @param null $entity
@@ -32,9 +25,9 @@ class FormService extends ServiceAbstract
         if (!$entity) {
             $entity = $this->getEntity($className);
         }
-        $formName = 'calendar_'.$entity->get('underscore_entity_name').'_form';
+        $formName = 'calendar_' . $entity->get('underscore_entity_name') . '_form';
         $form = $this->getServiceLocator()->get($formName);
-        $filterName = 'calendar_'.$entity->get('underscore_entity_name').'_form_filter';
+        $filterName = 'calendar_' . $entity->get('underscore_entity_name') . '_form_filter';
         $filter = $this->getServiceLocator()->get($filterName);
         $form->setInputFilter($filter);
         if ($bind) {
