@@ -6,7 +6,7 @@
  * @category    Calendar
  *
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2014 ITEA Office (http://itea3.org)
+ * @copyright   Copyright (c) 2004-2014 ITEA Office (https://itea3.org)
  */
 
 namespace Calendar\View\Helper;
@@ -14,6 +14,7 @@ namespace Calendar\View\Helper;
 use Calendar\Acl\Assertion\Calendar as CalendarAssertion;
 use Calendar\Entity\Calendar;
 use Calendar\Service\CalendarService;
+use Project\Entity\Project;
 
 /**
  * Create a link to an calendar.
@@ -23,25 +24,13 @@ use Calendar\Service\CalendarService;
 class CalendarLink extends LinkAbstract
 {
     /**
-     * @var Calendar
-     */
-    protected $calendar;
-    /**
-     * @var int
-     */
-    protected $year;
-    /**
-     * @var
-     */
-    protected $which;
-
-    /**
      * @param Calendar $calendar
      * @param string   $action
      * @param string   $show
      * @param string   $which
      * @param null     $alternativeShow
      * @param null     $year
+     * @param Project  $project
      *
      * @return string
      *
@@ -53,41 +42,36 @@ class CalendarLink extends LinkAbstract
         $show = 'name',
         $which = CalendarService::WHICH_UPCOMING,
         $alternativeShow = null,
-        $year = null
+        $year = null,
+        Project $project = null
     ) {
         $this->setCalendar($calendar);
         $this->setAction($action);
         $this->setShow($show);
         $this->setWhich($which);
         $this->setYear($year);
+        $this->setProject($project);
         $this->setAlternativeShow($alternativeShow);
-        /**LiLik
+        /**
          * Set the non-standard options needed to give an other link value
          */
-        $this->setShowOptions(
-            [
-                'alternativeShow' => $this->getAlternativeShow(),
-                'text-which-tab'  => ucfirst($this->getWhich()),
-                'name'            => $this->getCalendar()->getCalendar(),
-            ]
-        );
+        $this->setShowOptions([
+            'alternativeShow' => $this->getAlternativeShow(),
+            'text-which-tab'  => ucfirst($this->getWhich()),
+            'name'            => $this->getCalendar()->getCalendar(),
+        ]);
 
         /*
          * Check the access to the object
          */
-        if (!$this->hasAccess(
-            $this->getCalendar(),
-            CalendarAssertion::class,
-            $this->getAction()
-        )
-        ) {
+        if (!$this->hasAccess($this->getCalendar(), CalendarAssertion::class, $this->getAction())) {
             return '';
         }
 
-        $this->addRouterParam('entity', 'calendar');
         $this->addRouterParam('id', $this->getCalendar()->getId());
         $this->addRouterParam('calendar', $this->getCalendar()->getId());
         $this->addRouterParam('docRef', $this->getCalendar()->getDocRef());
+        $this->addRouterParam('project', $this->getProject()->getId());
         $this->addRouterParam('which', $this->getWhich());
         $this->addRouterParam('year', $this->getYear());
 
@@ -144,6 +128,10 @@ class CalendarLink extends LinkAbstract
                 $this->setRouter('community/calendar/send-message');
                 $this->setText($this->translate("txt-send-message-to-attendees"));
                 break;
+            case 'download-binder':
+                $this->setRouter('community/calendar/download-binder');
+                $this->setText($this->translate("txt-download-binder"));
+                break;
             case 'presence-list':
                 $this->setRouter('community/calendar/presence-list');
                 $this->setText($this->translate("txt-download-presence-list"));
@@ -153,12 +141,13 @@ class CalendarLink extends LinkAbstract
                 $this->setText(sprintf($this->translate("txt-view-calendar-%s"), $this->getCalendar()));
                 break;
             case 'view':
-                $this->setRouter('route-'.$this->getCalendar()->get("underscore_full_entity_name"));
+                $this->setRouter('route-' . $this->getCalendar()->get("underscore_full_entity_name"));
                 $this->addRouterParam('calendar', $this->getCalendar()->getId());
                 $this->addRouterParam('docRef', $this->getCalendar()->getDocRef());
-                $this->setText(
-                    sprintf($this->translate("txt-view-calendar-item-%s"), $this->getCalendar()->getCalendar())
-                );
+                $this->setText(sprintf(
+                    $this->translate("txt-view-calendar-item-%s"),
+                    $this->getCalendar()->getCalendar()
+                ));
                 break;
             case 'view-community':
                 $this->setRouter('community/calendar/calendar');
@@ -168,60 +157,21 @@ class CalendarLink extends LinkAbstract
                 $this->setRouter('zfcadmin/calendar-manager/calendar');
                 $this->setText(sprintf($this->translate("txt-view-calendar-%s"), $this->getCalendar()));
                 break;
+            case 'edit-attendees-admin':
+                $this->setRouter('zfcadmin/calendar-manager/select-attendees');
+                $this->setText(sprintf($this->translate("txt-select-attendees-for-calendar-%s"), $this->getCalendar()));
+                break;
+            case 'new':
+                $this->setRouter('zfcadmin/calendar-manager/new');
+                if (is_null($this->getProject())) {
+                    $this->setText(sprintf($this->translate("txt-add-calendar-item")));
+                } else {
+                    $this->setText(sprintf($this->translate("txt-review-meeting-for-%s"), $this->getProject()));
+                }
+
+                break;
             default:
                 throw new \Exception(sprintf("%s is an incorrect action for %s", $this->getAction(), __CLASS__));
         }
-    }
-
-    /**
-     * @return Calendar
-     */
-    public function getCalendar()
-    {
-        if (is_null($this->calendar)) {
-            $this->calendar = new Calendar();
-        }
-
-        return $this->calendar;
-    }
-
-    /**
-     * @param Calendar $calendar
-     */
-    public function setCalendar($calendar)
-    {
-        $this->calendar = $calendar;
-    }
-
-    /**
-     * @return int
-     */
-    public function getYear()
-    {
-        return $this->year;
-    }
-
-    /**
-     * @param int $year
-     */
-    public function setYear($year)
-    {
-        $this->year = $year;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getWhich()
-    {
-        return $this->which;
-    }
-
-    /**
-     * @param mixed $which
-     */
-    public function setWhich($which)
-    {
-        $this->which = $which;
     }
 }
