@@ -2,30 +2,31 @@
 // Generated by ZF2's ./bin/templatemap_generator.php
 return array(
     'calendar/calendar/calendar-type-color-css'     => __DIR__ . '//view/calendar/calendar/calendar-type-color-css.twig',
-    'calendar/calendar-manager/edit'                => __DIR__ . '//view/calendar/calendar-manager/edit.twig',
-    'calendar/calendar-manager/calendar'            => __DIR__ . '//view/calendar/calendar-manager/calendar.twig',
-    'calendar/calendar-manager/new'                 => __DIR__ . '//view/calendar/calendar-manager/new.twig',
-    'calendar/calendar-manager/overview'            => __DIR__ . '//view/calendar/calendar-manager/overview.twig',
-    'calendar/partial/form/calendar'                => __DIR__ . '//view/calendar/partial/form/calendar.twig',
     'calendar/partial/entity/calendar'              => __DIR__ . '//view/calendar/partial/entity/calendar.twig',
-    'calendar/partial/year-selector'                => __DIR__ . '//view/calendar/partial/year-selector.twig',
-    'calendar/partial/list/calendar-community'      => __DIR__ . '//view/calendar/partial/list/calendar-community.twig',
-    'calendar/partial/list/calendar'                => __DIR__ . '//view/calendar/partial/list/calendar.twig',
-    'calendar/partial/list/calendar-past'           => __DIR__ . '//view/calendar/partial/list/calendar-past.twig',
-    'calendar/partial/list/calendar-thumb'          => __DIR__ . '//view/calendar/partial/list/calendar-thumb.twig',
+    'calendar/partial/list/calendar-contact'        => __DIR__ . '//view/calendar/partial/list/calendar-contact.twig',
     'calendar/partial/list/calendar-small'          => __DIR__ . '//view/calendar/partial/list/calendar-small.twig',
     'calendar/partial/list/calendar-upcoming-thumb' => __DIR__ . '//view/calendar/partial/list/calendar-upcoming-thumb.twig',
+    'calendar/partial/list/calendar-thumb'          => __DIR__ . '//view/calendar/partial/list/calendar-thumb.twig',
+    'calendar/partial/list/calendar-community'      => __DIR__ . '//view/calendar/partial/list/calendar-community.twig',
+    'calendar/partial/list/calendar-past'           => __DIR__ . '//view/calendar/partial/list/calendar-past.twig',
+    'calendar/partial/list/calendar'                => __DIR__ . '//view/calendar/partial/list/calendar.twig',
     'calendar/partial/list/calendar-past-thumb'     => __DIR__ . '//view/calendar/partial/list/calendar-past-thumb.twig',
-    'calendar/partial/list/calendar-contact'        => __DIR__ . '//view/calendar/partial/list/calendar-contact.twig',
     'calendar/partial/pagination-control'           => __DIR__ . '//view/calendar/partial/pagination-control.twig',
+    'calendar/partial/form/calendar'                => __DIR__ . '//view/calendar/partial/form/calendar.twig',
+    'calendar/partial/year-selector'                => __DIR__ . '//view/calendar/partial/year-selector.twig',
     'calendar/calendar-document/document'           => __DIR__ . '//view/calendar/calendar-document/document.twig',
     'calendar/calendar-document/edit'               => __DIR__ . '//view/calendar/calendar-document/edit.twig',
-    'calendar/pdf/review-calendar'                  => __DIR__ . '//view/calendar/pdf/review-calendar.twig',
     'calendar/pdf/calendar-contact'                 => __DIR__ . '//view/calendar/pdf/calendar-contact.twig',
-    'calendar/calendar-community/contact'           => __DIR__ . '//view/calendar/calendar-community/contact.twig',
+    'calendar/pdf/review-calendar'                  => __DIR__ . '//view/calendar/pdf/review-calendar.twig',
     'calendar/calendar-community/send-message'      => __DIR__ . '//view/calendar/calendar-community/send-message.twig',
-    'calendar/calendar-community/calendar'          => __DIR__ . '//view/calendar/calendar-community/calendar.twig',
-    'calendar/calendar-community/select-attendees'  => __DIR__ . '//view/calendar/calendar-community/select-attendees.twig',
     'calendar/calendar-community/overview'          => __DIR__ . '//view/calendar/calendar-community/overview.twig',
     'calendar/calendar-community/review-calendar'   => __DIR__ . '//view/calendar/calendar-community/review-calendar.twig',
+    'calendar/calendar-community/select-attendees'  => __DIR__ . '//view/calendar/calendar-community/select-attendees.twig',
+    'calendar/calendar-community/calendar'          => __DIR__ . '//view/calendar/calendar-community/calendar.twig',
+    'calendar/calendar-community/contact'           => __DIR__ . '//view/calendar/calendar-community/contact.twig',
+    'calendar/calendar-manager/overview'            => __DIR__ . '//view/calendar/calendar-manager/overview.twig',
+    'calendar/calendar-manager/new'                 => __DIR__ . '//view/calendar/calendar-manager/new.twig',
+    'calendar/calendar-manager/select-attendees'    => __DIR__ . '//view/calendar/calendar-manager/select-attendees.twig',
+    'calendar/calendar-manager/calendar'            => __DIR__ . '//view/calendar/calendar-manager/calendar.twig',
+    'calendar/calendar-manager/edit'                => __DIR__ . '//view/calendar/calendar-manager/edit.twig',
 );
